Split native engine lifecycle, supervision and turn ownership
Preserve engine and callback behavior while extracting cohesive private modules for engine handles, worker lifecycle, supervision, turn execution, interaction and request dispatch.
