+++ packages/agent-native-bridge/ffi/Agent/CLI/MacOS/EngineEvents.hs
@@ -1,0 +1,73 @@
+{-# LANGUAGE ForeignFunctionInterface #-}
+
+-- | Native engine event envelopes and synchronous callback marshalling.
+-- Byte buffers remain alive until the client callback returns.
+module Agent.CLI.MacOS.EngineEvents
+    ( EventCallback
+    , boundedEventText
+    , successEvent
+    , failureEvent
+    , sendEvent
+    , sendBinaryEvent
+    ) where
+
+import Control.Exception.Safe (tryAny)
+import Control.Monad (void)
+import qualified Data.Aeson as Aeson
+import qualified Data.ByteString as BS
+import qualified Data.ByteString.Lazy as LBS
+import Data.Text (Text)
+import qualified Data.Text as Text
+import Data.Word (Word8)
+import Foreign (FunPtr, Ptr, castPtr)
+import Foreign.C.Types (CSize(..))
+
+type EventCallback = Ptr () -> Ptr Word8 -> CSize -> IO ()
+
+foreign import ccall "dynamic"
+    invokeEventCallback :: FunPtr EventCallback -> EventCallback
+
+boundedEventText :: Text -> (Text, Bool)
+boundedEventText value =
+    let (visible, remainder) = Text.splitAt 8192 value
+    in (visible, not (Text.null remainder))
+
+successEvent :: Aeson.ToJSON value => Text -> value -> Aeson.Value
+successEvent requestId result =
+    Aeson.object
+        [ "id" Aeson..= requestId
+        , "ok" Aeson..= True
+        , "result" Aeson..= result
+        ]
+
+failureEvent :: Text -> Text -> Aeson.Value
+failureEvent requestId message =
+    Aeson.object
+        [ "id" Aeson..= requestId
+        , "ok" Aeson..= False
+        , "error" Aeson..= message
+        ]
+
+sendEvent :: FunPtr EventCallback -> Ptr () -> Aeson.Value -> IO ()
+sendEvent callback context event =
+    void $ tryAny $
+        BS.useAsCStringLen (LBS.toStrict (Aeson.encode event)) \(bytes, length) ->
+            sendCallbackBytes callback context (castPtr bytes) length
+
+sendBinaryEvent :: FunPtr EventCallback -> Ptr () -> BS.ByteString -> IO ()
+sendBinaryEvent callback context bytes =
+    void $ tryAny $
+        BS.useAsCStringLen bytes \ (pointer, length) ->
+            sendCallbackBytes callback context (castPtr pointer) length
+
+sendCallbackBytes
+    :: FunPtr EventCallback
+    -> Ptr ()
+    -> Ptr Word8
+    -> Int
+    -> IO ()
+sendCallbackBytes callback context bytes length =
+    invokeEventCallback callback
+        context
+        (castPtr bytes)
+        (fromIntegral length)
